\name{IntroR-package}
\alias{IntroR-package}
\alias{IntroR}
\docType{package}
\title{
Introductory Statistics and Data Analysis Using R
}
\description{
Statistics and Data Analysis Using R provides a brief but systematic introduction to the R statistical software suite for biomedical scientists.  The primary objective is learning to use R, but the course also emphasizes the standards of practice that programmers and data analysts have implemented to ensure transparency, accuracy and accountability. No formal background in statistics or programming is necessary.
}
\details{
\tabular{ll}{
Package: \tab IntroR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-11\cr
License: \tab What license is it under?\cr
}
}
\author{
  LC, EF, LM, and RS
}

\keyword{ package }
\examples{
## see vignettes
}
