% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introR.R
\docType{package}
\name{introR}
\alias{introR}
\title{\code{introR} package}
\description{
Aprenda R con ejercicios interactivos del curso introducción a R de ixpantia
}
\details{
La documentación es parte de este paquete, por favor usar el eslabón `index`
que aparece aqui abajo.
}
