% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixpantia-introR-package.R
\docType{package}
\name{ixpantia.introR-package}
\alias{ixpantia.introR}
\alias{ixpantia.introR-package}
\title{Paquete en R para incluir ejercicios en RStudio}
\description{
El paquete ixpantia.introR da acceso a una serie de ejercicios dentro de la
interfaz de RStudio.
Ofreemos los siguientes ejercicios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ixpantia/ixpantia.introR}
  \item Report bugs at \url{https://github.com/ixpantia/ixpantia.introR/issues}
}

}
\author{
\strong{Maintainer}: Frans van Dunné \email{frans@ixpantia.com} (\href{https://orcid.org/0000-0002-7853-2811}{ORCID})

Authors:
\itemize{
  \item Andrea Vargas \email{andrea@ixpantia.com}
  \item Ronny Hernández Mora \email{ronny@ixpantia.com} (\href{https://orcid.org/0000-0001-6225-7096}{ORCID})
  \item Magdiel Ablan \email{magdiel@ixpantia.com}
}

Other contributors:
\itemize{
  \item ixpantia, 
         family = SRL \email{hola@ixpantia.com} [copyright holder]
}

}
\keyword{internal}
