% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixpantia.introR.R
\docType{package}
\name{ixpantia.introR}
\alias{ixpantia.introR}
\title{\code{ixpantia.introR} package}
\description{
Aprenda R con ejercicios interactivos del curso introducción a R de ixpantia
}
\details{
La documentación es parte de este paquete, por favor usar el eslabón `index`
que aparece aqui abajo.
}
