% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_tutoriales.R
\name{lista_tutoriales}
\alias{lista_tutoriales}
\title{Función para listar los tutoriales disponibles en el paquete}
\usage{
lista_tutoriales()
}
\value{
lista Lista de tutoriales
}
\description{
Función para listar los tutoriales disponibles en el paquete
}
