% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_tutoriales.R
\name{listar_tutoriales}
\alias{listar_tutoriales}
\title{Listar tutoriales}
\usage{
listar_tutoriales(tipo)
}
\arguments{
\item{tipo}{Es el tipo del tutorial que queremos listar. Puede ser `tarea` o
`pizarra`}
}
\description{
Genera una lista con los tutoriales disponibles en el paquete.
}
\details{
La función permitar dar una lista de los tutoriales disponibles en
el paquete por su tipo, ya sea tarea o pizarra.
}
\examples{
\dontrun{
listar_tutoriales(tipo = "pizarra")
}
}
