% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_tutoriales.R
\name{practicar}
\alias{practicar}
\title{Practicar con tutorial}
\usage{
practicar(nombre_tutorial)
}
\arguments{
\item{nombre_tutorial}{Es el nombre de uno de los tutoriales dentro del
paquete ixpantia.introR}
}
\description{
Despliega el tutorial elegido por su nombre en un navegador web
}
\details{
La función necesita que en el computador docker esté instalado y que
se hayan agregado las lineas siguientes en la terminal:
}
\examples{
\dontrun{
practicar("pizarra_coercion")
}
}
