% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lista_tutoriales.R
\name{revisar_dependencias}
\alias{revisar_dependencias}
\title{Instalar dependendencias}
\usage{
revisar_dependencias(instalar = FALSE)
}
\arguments{
\item{instalar}{Indicar si quieres instalar las dependencias al revisar}
}
\description{
Muestra las dependencias del paquete introR que son
necesarias para poder practicar con los tutoriales interactivos
}
\details{
El paquete introR toma en cuenta una cantidad de paquetes
que son necesarios para practicar con los tutoriales interactivos. Sino
tenemos estos paquetes instalados en nuestro computador, no podremos hacer
uso de los tutoriales. Si en algún caso no sabemos cuál paquete nos falta,
esta función nos dará la lista de dependencias.
}
\examples{
\dontrun{
listar_tutoriales(tipo = "pizarra")
}
}
